// frontend/src/app/api/models/source-info.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CollectionInfo } from '../models/collection-info';
import { IncludeInfo } from '../models/include-info';
import { SourceReport } from '../models/source-report';
export interface SourceInfo {
  collection: string;
  excludeDirs: Array<string>;
  excludeNames: Array<string>;
  includes: Array<IncludeInfo>;
  lastRun?: (SourceReport | null);
  path: string;
  pathExists: boolean;
  secretAllow: Array<string>;
  stats?: (CollectionInfo | null);
  title: string;
}